[MB-1670] Use autopilot to call takeoff
