+++ android-plugin-lib/src/main/AndroidManifest.xml
@@ -17,6 +17,8 @@
             </intent-filter>
         </receiver>
 
+        <meta-data android:name="com.urbanairship.autopilot" android:value="com.urbanairship.unityplugin.UnityAutopilot" />
+
     </application>
    
 </manifest>
